requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128 https://download.pytorch.org/whl/cu128
torch==2.9.0+cu128
torchvision==0.24.0+cu128
torchaudio==2.9.0+cu128
triton==3.5.0

transformers==4.57.3
diffusers==0.36.0
accelerate==1.6.0
sentencepiece==0.2.1
safetensors==0.7.0
tokenizers==0.22.2
einops==0.8.1
xgrammar==0.1.19
outlines==0.1.11
lm-format-enforcer==0.10.12

opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
open3d==0.19.0
imageio==2.34.0
imageio-ffmpeg==0.5.1
kornia==0.8.2
pillow==12.0.0
matplotlib==3.10.8
scikit-image==0.25.2
plyfile==0.8.1
pyquaternion==0.9.9
cupy-cuda12x==13.6.0

numpy==2.2.6
pandas==2.3.3
scipy==1.15.3
scikit-learn==1.7.2
numba==0.61.2

fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
httpx==0.28.1
websockets==15.0.1
flask==3.1.2
flask-cors==6.0.2
rclpy
cv-bridge
launch
launch-ros

tqdm==4.67.1
rich==14.2.0
ipython==8.38.0
jupyter-core==5.9.1
ipywidgets==8.1.8
ray==2.53.0
omegaconf==2.3.0
commentjson==0.9.0
configargparse==1.7.1
addict==2.4.0
pyyaml==6.0.3
huggingface-hub==0.36.0
datasets
evaluate

oss2==2.19.1
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-exporter-otlp==1.39.1
sentry-sdk==2.48.0

dash==3.3.0
plotly==6.5.0
kaleido==1.2.0
moviepy==1.0.3
openai==2.14.0
qwen-vl-utils==0.0.14
mistral-common==1.8.8
tiktoken==0.12.0
e3nn==0.5.3
